# docs/swagger.yaml
basePath: /
definitions:
...
  fortyfour-backend_internal_dto.ErrorResponse:
    properties:
      message:
        example: invalid credentials
        type: string
    type: object
...
  fortyfour-backend_internal_dto.SektorResponse:
    properties:
      created_at:
        type: string
      id:
        type: string
      nama_sektor:
        type: string
      sub_sektor:
        items:
          $ref: '#/definitions/fortyfour-backend_internal_dto.SubSektorResponse'
        type: array
      updated_at:
        type: string
    type: object
...
  fortyfour-backend_internal_dto.SubSektorResponse:
    properties:
      created_at:
        type: string
      id:
        type: string
      id_sektor:
        type: string
      nama_sektor:
        type: string
      nama_sub_sektor:
        type: string
      updated_at:
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: API documentation for Fortyfour Backend - main auth and management
    service.
  title: Fortyfour Backend API
  version: "1.0"
paths:
...
  /api/sektor:
    get:
      description: Mengambil seluruh data sektor beserta sub sektor
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/fortyfour-backend_internal_dto.SektorResponse'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/fortyfour-backend_internal_dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List semua sektor
      tags:
      - Sektor
  /api/sektor/{id}:
    get:
      description: Mengambil satu data sektor beserta sub sektor
      parameters:
      - description: Sektor ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/fortyfour-backend_internal_dto.SektorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/fortyfour-backend_internal_dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Ambil sektor berdasarkan ID
      tags:
      - Sektor
...
  /api/sub_sektor:
    get:
      description: Mengambil seluruh data sub sektor
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/fortyfour-backend_internal_dto.SubSektorResponse'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/fortyfour-backend_internal_dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List semua sub sektor
      tags:
      - SubSektor
  /api/sub_sektor/{id}:
    get:
      description: Mengambil satu data sub sektor
      parameters:
      - description: SubSektor ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/fortyfour-backend_internal_dto.SubSektorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/fortyfour-backend_internal_dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Ambil sub sektor berdasarkan ID
      tags:
      - SubSektor
  /api/sub_sektor/by_sektor/{id}:
    get:
      description: Mengambil data sub sektor dalam satu sektor
      parameters:
      - description: Sektor ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/fortyfour-backend_internal_dto.SubSektorResponse'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/fortyfour-backend_internal_dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Ambil sub sektor berdasarkan Sektor ID
      tags:
      - SubSektor
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"